FEAT: PS240223

+++ src/PS240222/BJ_JH1654.java
@@ -7,8 +7,17 @@ import java.util.Arrays;
 import java.util.StringTokenizer;
 
 public class BJ_JH1654 {
+	/*
+	 * 1. 최댓값을 기반으로 이진 탐색을 진행한다.
+	 * 2. 조각이 너무 작으면(N보다 total이 크게 나오면) , 더 오른쪽에서 찾는다. 이때의 최댓값을 저장해둔다.
+	 * 3. 조각이 너무 크면(N보다 total이 작게 나오면), 더 왼쪽에서 찾는다.
+	 * 4. 이 과정에서 start가 end보다 커지면, 현재 mid부터 쭉 올라가면서 N보다 total이 크거나 같을 때까지
+	 * 올라간다.
+	 * 
+	 * 
+	 */
 	static long[] lan;
-	static int N,K;
+	static long N,K;
 	static long mid, total;
 	
 	public static void main(String[] args) throws IOException {
@@ -21,7 +30,7 @@ public class BJ_JH1654 {
 		lan = new long[K];
 		
 		for (int i = 0; i < K; i++) {
-			int temp = Integer.parseInt(br.readLine());
+			long temp = Integer.parseInt(br.readLine());
 			
 			lan[i] = temp;
 		}
@@ -31,13 +40,13 @@ public class BJ_JH1654 {
 		//최댓값
 		long temp = lan[K-1];
 		
-		long[] bs = new long[(int) temp];
+		long[] bs = new long[temp];
 		
 		for (int i = 0; i < bs.length; i++) {
 			bs[i] = i+1;
 		}
 		
-		b_search(bs, 0,temp-1);
+		b_search(bs, 0, temp-1);
 	}
 	
 	//target이 딱히 존재하진 않고, 다른 것들 조각 다 더한게
